+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/model/AccountSettingsUserMFA.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2022.
+ * (C) Copyright IBM Corp. 2023.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -23,6 +23,7 @@ public class AccountSettingsUserMFA extends GenericModel {
   /**
    * Defines the MFA requirement for the user. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -32,6 +33,8 @@ public class AccountSettingsUserMFA extends GenericModel {
   public interface Mfa {
     /** NONE. */
     String NONE = "NONE";
+    /** NONE_NO_ROPC. */
+    String NONE_NO_ROPC = "NONE_NO_ROPC";
     /** TOTP. */
     String TOTP = "TOTP";
     /** TOTP4ALL. */
@@ -150,6 +153,7 @@ public class AccountSettingsUserMFA extends GenericModel {
    *
    * Defines the MFA requirement for the user. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users