+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/model/IdBasedMfaEnrollment.java
@@ -23,6 +23,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
   /**
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -32,6 +33,8 @@ public class IdBasedMfaEnrollment extends GenericModel {
   public interface TraitAccountDefault {
     /** NONE. */
     String NONE = "NONE";
+    /** NONE_NO_ROPC. */
+    String NONE_NO_ROPC = "NONE_NO_ROPC";
     /** TOTP. */
     String TOTP = "TOTP";
     /** TOTP4ALL. */
@@ -47,6 +50,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
   /**
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -56,6 +60,8 @@ public class IdBasedMfaEnrollment extends GenericModel {
   public interface TraitUserSpecific {
     /** NONE. */
     String NONE = "NONE";
+    /** NONE_NO_ROPC. */
+    String NONE_NO_ROPC = "NONE_NO_ROPC";
     /** TOTP. */
     String TOTP = "TOTP";
     /** TOTP4ALL. */
@@ -71,6 +77,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
   /**
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -80,6 +87,8 @@ public class IdBasedMfaEnrollment extends GenericModel {
   public interface TraitEffective {
     /** NONE. */
     String NONE = "NONE";
+    /** NONE_NO_ROPC. */
+    String NONE_NO_ROPC = "NONE_NO_ROPC";
     /** TOTP. */
     String TOTP = "TOTP";
     /** TOTP4ALL. */
@@ -107,6 +116,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
    *
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -124,6 +134,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
    *
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users
@@ -141,6 +152,7 @@ public class IdBasedMfaEnrollment extends GenericModel {
    *
    * Defines the MFA trait for the account. Valid values:
    *   * NONE - No MFA trait set
+   *   * NONE_NO_ROPC- No MFA, disable CLI logins with only a password
    *   * TOTP - For all non-federated IBMId users
    *   * TOTP4ALL - For all users
    *   * LEVEL1 - Email-based MFA for all users