+++ modules/iam-identity/src/main/java/com/ibm/cloud/platform_services/iam_identity/v1/model/ProfileLink.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2022.
+ * (C) Copyright IBM Corp. 2023.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -39,7 +39,7 @@ public class ProfileLink extends GenericModel {
   /**
    * Gets the id.
    *
-   * the unique identifier of the claim rule.
+   * the unique identifier of the link.
    *
    * @return the id
    */
@@ -50,7 +50,7 @@ public class ProfileLink extends GenericModel {
   /**
    * Gets the entityTag.
    *
-   * version of the claim rule.
+   * version of the link.
    *
    * @return the entityTag
    */